Operational.

+++ _prog/core.sh
@@ -1,11 +1,16 @@
 ##### Core
 
-resetPulse() {
+_resetPulse() {
+	pgrep kmix >/dev/null 2>&1 && resetKmix="true"
+	
 	pkill pulse
 	pax11publish -r
 	start-pulseaudio-x11
 	sleep 1
-	pkill kmix
-	kmix >/dev/null 2>&1 &
 	
+	if [[ "$resetKmix" == "true" ]]
+	then
+		pkill kmix
+		kmix >/dev/null 2>&1 &
+	fi
 }
